requirements.txt
Flask==3.0.3
joblib==1.4.0
flask-cors
numpy~=1.26.4
google-cloud-vision